pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "weavelens"
dynamic = ["version"]
description = "Local-first RAG service with Weaviate v4, FastAPI, and Telegram bot"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [
    {name = "Michael", email = "onisoris@yandex.ru"},
    {name = "Egor", email = "egor.goncharov.oisa@mail.ru"}
]
urls = {Homepage = "https://github.com/OnisOris/weavelens"}

dependencies = [
  "weaviate-client>=4.16.9",
  "fastapi",
  "uvicorn[standard]>=0.30",
  "aiogram",
  "pydantic",
  "python-dotenv>=1.0",
  "httpx>=0.27",
  "watchfiles>=0.21",
  "sentence-transformers>=3.0",
  "numpy>=1.26",
  "scikit-learn>=1.5",
  "prometheus-client>=0.20",
  "orjson>=3.10",
  "cryptography>=43.0",
  "pypdf>=4.2",
  "python-docx>=1.1.2"
]

[project.optional-dependencies]
llm = ["ollama>=0.3"]
dev = ["ruff>=0.5", "mypy>=1.11", "pytest>=8.2", "pytest-asyncio>=0.23"]

[project.scripts]
weavelens-api = "weavelens.api.main:run"
weavelens-bot = "weavelens.bot.tg_bot:run"
weavelens-index = "weavelens.pipeline.index:cli"

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "pytest>=8.4.1",
    "ruff>=0.12.11",
]

[tool.ruff]
line-length = 100
select = [
"E",
"F",
"I",
"UP"
]
